refactor(cli): move AppConfig to models::config to break mega-SCC (T-SP-1 #7)
Move the `AppConfig` struct, its impls (`from_file`, `from_source`,
`from_content`, `validate`, `default_config_content`, `create_default_config`,
`needs_first_run`, `default_path`, `resolve_env_vars`) and the related
helpers (`find_project_config`, `merge_project_config`) from
`src/cli/mod.rs` (+ `src/cli/defaults.rs` + `src/cli/validation.rs`) into
a new leaf module `src/models/config.rs`.

Sub-config structs (`ServerConfig`, `RouterConfig`, `ProviderConfig`, ...)
remain in `src/cli/config.rs` and are re-imported by `models::config`.
The reverse-direction edge (`models::config -> cli`) replaces the previous
fan-in (`cli -> features/auth/...`), removing 5+ heavy SCC edges that
made `cli` part of the mega-SCC.

`crate::cli::AppConfig` keeps working via a re-export from
`crate::models::config` so external call sites compile unchanged. Internal
callers that imported via `use crate::cli::AppConfig` are switched to the
canonical `use crate::models::config::AppConfig`.

Effect on dependency graph:
- `cli/mod.rs` no longer pulls `auth::jwt`, `features::dlp`,
  `features::mcp`, `features::pledge`, `features::tap`,
  `features::policies`, `features::log_export` (via AppConfig fields).
- `models::config` is the single fan-in point for sub-configs, isolating
  the cycle to one module instead of spreading it across the graph.

Tests: 949 lib tests pass, `cargo build --all-targets` green,
`cargo clippy --all-targets -- -D warnings` clean,
`RUSTFLAGS=-D warnings cargo check --no-default-features` clean.
