+++ src/server/config_guard.rs
@@ -65,7 +65,7 @@ pub fn is_key_denied(section: &ConfigSection, key: &str) -> bool {
 /// hot-reload (re-parse + provider rebuild) fails.
 pub async fn persist_and_reload(
     state: &Arc<super::AppState>,
-    config: &crate::cli::AppConfig,
+    config: &crate::models::config::AppConfig,
 ) -> Result<(), super::AppError> {
     let config_path = match &state.config_source {
         crate::cli::ConfigSource::File(p) => p,
@@ -99,7 +99,7 @@ pub async fn persist_and_reload(
 /// Rebuilds [`ReloadableState`] from a validated config and atomically swaps it.
 fn reload_state(
     state: &Arc<super::AppState>,
-    config: crate::cli::AppConfig,
+    config: crate::models::config::AppConfig,
     _config_path: &Path,
 ) -> Result<(), super::AppError> {
     let new_router = crate::routing::classify::Router::new(config.clone());
