+++ src/server/lifecycle.rs
@@ -5,7 +5,7 @@
 //! `rustls-pki-types`.
 
 use super::{oauth_handlers, AppState};
-use crate::cli::AppConfig;
+use crate::models::config::AppConfig;
 use axum::{routing::get, Router as AxumRouter};
 use std::sync::Arc;
 use tokio::net::TcpListener;
