+++ src/server/mod.rs
@@ -55,9 +55,9 @@ pub(crate) use middleware::{
 };
 
 use crate::auth::TokenStore;
-use crate::cli::AppConfig;
 use crate::features::dlp::session::DlpSessionManager;
 use crate::features::token_pricing::SharedPricingTable;
+use crate::models::config::AppConfig;
 use crate::providers::ProviderRegistry;
 use crate::routing::classify::Router;
 use crate::security::{AuditLog, RateLimiter};
