+++ src/server/rpc/server_ns.rs
@@ -44,7 +44,7 @@ pub async fn reload_config(
 ) -> Result<StatusResponse, ErrorObjectOwned> {
     require_role(caller, Role::Operator)?;
 
-    use crate::cli::AppConfig;
+    use crate::models::config::AppConfig;
     use crate::providers::ProviderRegistry;
     use crate::routing::classify::Router;
     use crate::server::ReloadableState;
